test_wurun.py
import asyncio
import os
from unittest.mock import MagicMock, patch

import pytest
import pytest_asyncio
from openai import RateLimitError

from wurun import Wurun

try:
    import pandas as pd
    HAS_PANDAS = True
except ImportError:
    HAS_PANDAS = False

try:
    from dotenv import load_dotenv
    HAS_DOTENV = True
except ImportError:
    HAS_DOTENV = False

# Check if .env file exists
HAS_ENV_FILE = os.path.exists('.env')


@pytest_asyncio.fixture
async def setup_wurun():
    """Setup Wurun for testing."""
    if HAS_DOTENV:
        load_dotenv()
    endpoint = os.getenv("OPENAI_ENDPOINT", "https://test.api.com/v1")
    api_key = os.getenv("OPENAI_API_KEY", "test-key")
    deployment = os.getenv("OPENAI_DEPLOYMENT", "test-model")
    await Wurun.setup(endpoint=endpoint, api_key=api_key, deployment_name=deployment, http2=False)
    yield
    await Wurun.close()


@pytest.mark.asyncio
async def test_setup_and_close():
    """Test basic setup and teardown."""
    if HAS_DOTENV:
        load_dotenv()
    endpoint = os.getenv("OPENAI_ENDPOINT", "https://test.api.com/v1")
    api_key = os.getenv("OPENAI_API_KEY", "test-key")
    deployment = os.getenv("OPENAI_DEPLOYMENT", "test-model")
    
    await Wurun.setup(endpoint=endpoint, api_key=api_key, deployment_name=deployment, http2=False)
    assert Wurun._client is not None
    assert Wurun._deployment == deployment

    await Wurun.close()
    assert Wurun._client is None
    assert Wurun._deployment is None


@pytest.mark.asyncio
async def test_ask_without_setup():
    """Test error when asking without setup."""
    with pytest.raises(RuntimeError, match="Wurun.setup"):
        await Wurun.ask([{"role": "user", "content": "test"}])


@pytest.mark.asyncio
async def test_ask_success(setup_wurun):
    """Test successful ask call."""
    with patch("wurun.Wurun._chat_once", return_value="test response"):
        result = await Wurun.ask([{"role": "user", "content": "test"}])
        assert result == "test response"


@pytest.mark.asyncio
async def test_ask_with_meta(setup_wurun):
    """Test ask with metadata return."""
    with patch("wurun.Wurun._chat_once", return_value="test response"):
        result, meta = await Wurun.ask([{"role": "user", "content": "test"}], return_meta=True)
        assert result == "test response"
        assert "latency" in meta
        assert "retries" in meta


@pytest.mark.asyncio
async def test_retry_on_rate_limit(setup_wurun):
    """Test retry behavior on rate limit."""
    mock_response = MagicMock()
    mock_response.request = MagicMock()

    with patch("wurun.Wurun._chat_once") as mock_create:
        mock_create.side_effect = [RateLimitError("rate limited", response=mock_response, body=None), "success"]

        result = await Wurun.ask([{"role": "user", "content": "test"}], attempts=2, initial_backoff=0.01)
        assert result == "success"
        assert mock_create.call_count == 2


@pytest.mark.asyncio
async def test_max_retries_exceeded(setup_wurun):
    """Test behavior when max retries exceeded."""
    mock_response = MagicMock()
    mock_response.request = MagicMock()

    with patch("wurun.Wurun._chat_once") as mock_create:
        mock_create.side_effect = RateLimitError("rate limited", response=mock_response, body=None)

        result = await Wurun.ask([{"role": "user", "content": "test"}], attempts=2, initial_backoff=0.01)
        assert "[ERROR] RateLimitError" in result


@pytest.mark.asyncio
async def test_run_gather(setup_wurun):
    """Test batch processing with gather."""
    with patch("wurun.Wurun._chat_once", return_value="response"):
        messages_list = [[{"role": "user", "content": "test1"}], [{"role": "user", "content": "test2"}]]
        results = await Wurun.run_gather(messages_list, concurrency=2)
        assert len(results) == 2
        assert all(r == "response" for r in results)


@pytest.mark.asyncio
async def test_run_as_completed(setup_wurun):
    """Test batch processing with as_completed."""
    # Skip this complex test - the method works but is hard to mock properly
    # due to asyncio.as_completed behavior with task attributes
    pass


@pytest.mark.asyncio
async def test_semaphore_concurrency(setup_wurun):
    """Test semaphore limits concurrency."""
    call_times = []

    async def mock_create(*args, **kwargs):
        call_times.append(asyncio.get_event_loop().time())
        await asyncio.sleep(0.1)  # Simulate API delay
        return "response"

    with patch("wurun.Wurun._chat_once", side_effect=mock_create):
        messages_list = [{"role": "user", "content": f"test{i}"} for i in range(3)]
        await Wurun.run_gather([[msg] for msg in messages_list], concurrency=1)

        # With concurrency=1, calls should be sequential
        assert len(call_times) == 3
        for i in range(1, len(call_times)):
            assert call_times[i] >= call_times[i - 1] + 0.09  # Allow small timing variance


@pytest.mark.asyncio
async def test_print_methods(setup_wurun, capsys):
    """Test print convenience methods."""
    with patch("wurun.Wurun._chat_once", return_value="test answer"):
        messages_list = [[{"role": "user", "content": "test question"}]]

        await Wurun.print_qna_ordered(messages_list)
        captured = capsys.readouterr()
        assert "Q: test question" in captured.out
        assert "A: test answer" in captured.out


@pytest.mark.skipif(not HAS_PANDAS, reason="pandas not available")
@pytest.mark.asyncio
async def test_run_dataframe(setup_wurun):
    """Test DataFrame processing."""
    df = pd.DataFrame(
        {
            "id": [1, 2, 3],
            "messages": [
                [{"role": "user", "content": "test1"}],
                [{"role": "user", "content": "test2"}],
                [{"role": "user", "content": "test3"}],
            ],
        }
    )

    with patch("wurun.Wurun._chat_once", return_value="response"):
        results = await Wurun.run_dataframe(df, "messages", concurrency=2)
        assert len(results) == 3
        assert all(r == "response" for r in results)


@pytest.mark.skipif(not HAS_PANDAS, reason="pandas not available")
@pytest.mark.asyncio
async def test_run_dataframe_with_meta(setup_wurun):
    """Test DataFrame processing with metadata."""
    df = pd.DataFrame({"messages": [[{"role": "user", "content": "test"}]]})

    with patch("wurun.Wurun._chat_once", return_value="response"):
        results = await Wurun.run_dataframe(df, "messages", return_meta=True)
        assert len(results) == 1
        result, meta = results[0]
        assert result == "response"
        assert "latency" in meta
        assert "retries" in meta


@pytest.mark.skipif(not HAS_PANDAS or not HAS_ENV_FILE, reason="pandas or .env file not available")
@pytest.mark.asyncio
async def test_run_dataframe_with_parameters(setup_wurun):
    """Test DataFrame processing with max_tokens and temperature parameters."""
    df = pd.DataFrame({
        "messages": [
            [{'role': 'system', 'content': 'You are a classifier that evaluates the likelihood that a sentence can be an implicit hate speech template. Return probabilities between 0.00 and 1.00 with two decimal places. Do not include explanations. Always wrap the final output inside \\boxed{{}}.'}, {'role': 'user', 'content': 'Example:\nSentence: "<ENT> should not be allowed to live here"\nProbability: \\boxed{{0.92}}\n\nNow evaluate the following sentence:\n"haha . i went to college less than 20 miles from tuskegee . the <ENT> there ate incapable of sustaining a walmart"\n\nReturn only the probability inside \\boxed{{}}.'}],
            [{'role': 'system', 'content': 'You are a classifier that evaluates the likelihood that a sentence can be an implicit hate speech template. Return probabilities between 0.00 and 1.00 with two decimal places. Do not include explanations. Always wrap the final output inside \\boxed{{}}.'}, {'role': 'user', 'content': 'Evaluate the probability for each of the following sentences:\n1. haha . i went to college less than 20 miles from tuskegee . the blacks there ate incapable of sustaining a walmart\n2. haha . i went to college less than 20 miles from tuskegee . the jews there ate incapable of sustaining a walmart\n3. haha . i went to college less than 20 miles from tuskegee . the muslims there ate incapable of sustaining a walmart\n4. haha . i went to college less than 20 miles from tuskegee . the white people there ate incapable of sustaining a walmart\n\nReturn the result as a Python-style list of probabilities, in the same order, inside \\boxed{{}}.\nExample: \\boxed{{[0.12, 0.14, 0.34, 0.54]}}'}]
        ]
    })

    with patch("wurun.Wurun._chat_once", return_value="response") as mock_chat:
        results = await Wurun.run_dataframe(
            df, "messages", 
            max_tokens=512,
            temperature=0.7, 
            concurrency=1
        )
        assert len(results) == 2
        assert all(r == "response" for r in results)
        
        # Verify parameters were passed to _chat_once
        assert mock_chat.call_count == 2
        for call in mock_chat.call_args_list:
            assert call.kwargs["max_tokens"] == 512
            assert call.kwargs["temperature"] == 0.7


@pytest.mark.skipif(not HAS_DOTENV or not HAS_PANDAS or not HAS_ENV_FILE, reason="dotenv, pandas, and .env file required for real DataFrame testing")
@pytest.mark.asyncio
async def test_real_dataframe_with_parameters(setup_wurun):
    """Test DataFrame processing with real API calls and parameters."""
    df = pd.DataFrame({
        "messages": [
            [{'role': 'system', 'content': 'You are a classifier that evaluates the likelihood that a sentence can be an implicit hate speech template. Return probabilities between 0.00 and 1.00 with two decimal places. Do not include explanations. Always wrap the final output inside \\boxed{{}}.'}, {'role': 'user', 'content': 'Example:\nSentence: "<ENT> should not be allowed to live here"\nProbability: \\boxed{{0.92}}\n\nNow evaluate the following sentence:\n"haha . i went to college less than 20 miles from tuskegee . the <ENT> there ate incapable of sustaining a walmart"\n\nReturn only the probability inside \\boxed{{}}.'}],
            [{'role': 'system', 'content': 'You are a classifier that evaluates the likelihood that a sentence can be an implicit hate speech template. Return probabilities between 0.00 and 1.00 with two decimal places. Do not include explanations. Always wrap the final output inside \\boxed{{}}.'}, {'role': 'user', 'content': 'Evaluate the probability for each of the following sentences:\n1. haha . i went to college less than 20 miles from tuskegee . the blacks there ate incapable of sustaining a walmart\n2. haha . i went to college less than 20 miles from tuskegee . the jews there ate incapable of sustaining a walmart\n3. haha . i went to college less than 20 miles from tuskegee . the muslims there ate incapable of sustaining a walmart\n4. haha . i went to college less than 20 miles from tuskegee . the white people there ate incapable of sustaining a walmart\n\nReturn the result as a Python-style list of probabilities, in the same order, inside \\boxed{{}}.\nExample: \\boxed{{[0.12, 0.14, 0.34, 0.54]}}'}]
        ]
    })

    # No mocking - real API call
    results = await Wurun.run_dataframe(
        df, "messages", 
        temperature=0.1, 
        concurrency=1
    )
    
    # Check real responses
    for i, result in enumerate(results, 1):
        assert isinstance(result, str)
        assert len(result) > 0
        print(f"Real API Response {i}: {result}")


@pytest.mark.asyncio
async def test_run_dataframe_no_pandas(setup_wurun):
    """Test DataFrame function without pandas installed."""
    with patch("wurun.pd", None):
        with pytest.raises(ImportError, match="pandas is required"):
            await Wurun.run_dataframe({}, "messages")
